Create directory for profiler options tests
