+++ dataprofiler/tests/profilers/profiler_options/test_base_option.py
@@ -1,7 +1,7 @@
 import unittest
 
 from dataprofiler.profilers.profiler_options import BaseOption
-from dataprofiler.tests.profilers.abstract_test_options import AbstractTestOptions
+from dataprofiler.tests.profilers.profiler_options.abstract_test_options import AbstractTestOptions
 
 
 class TestBaseOption(AbstractTestOptions, unittest.TestCase):
