+++ dataprofiler/tests/profilers/profiler_options/test_categorical_options.py
@@ -1,11 +1,11 @@
-from dataprofiler.profilers.profiler_options import DateTimeOptions
-from dataprofiler.tests.profilers.test_base_column_options import TestBaseColumnOptions
+from dataprofiler.profilers.profiler_options import CategoricalOptions
+from dataprofiler.tests.profilers.profiler_options.test_base_column_options import TestBaseColumnOptions
 
 
-class TestDateTimeOptions(TestBaseColumnOptions):
-
-    option_class = DateTimeOptions
-        
+class TestCategoricalOptions(TestBaseColumnOptions):
+    
+    option_class = CategoricalOptions
+    
     def test_init(self):
         super().test_init()
