+++ dataprofiler/tests/profilers/profiler_options/test_datalabeler_options.py
@@ -1,5 +1,5 @@
 from dataprofiler.profilers.profiler_options import DataLabelerOptions
-from dataprofiler.tests.profilers.test_base_column_options import TestBaseColumnOptions
+from dataprofiler.tests.profilers.profiler_options.test_base_column_options import TestBaseColumnOptions
 from dataprofiler.labelers.base_data_labeler import BaseDataLabeler
