+++ dataprofiler/tests/profilers/profiler_options/test_datetime_options.py
@@ -1,11 +1,11 @@
-from dataprofiler.profilers.profiler_options import OrderOptions
-from dataprofiler.tests.profilers.test_base_column_options import TestBaseColumnOptions
+from dataprofiler.profilers.profiler_options import DateTimeOptions
+from dataprofiler.tests.profilers.profiler_options.test_base_column_options import TestBaseColumnOptions
 
 
-class TestOrderOptions(TestBaseColumnOptions):
-    
-    options_class = OrderOptions
-    
+class TestDateTimeOptions(TestBaseColumnOptions):
+
+    option_class = DateTimeOptions
+        
     def test_init(self):
         super().test_init()
