+++ dataprofiler/tests/profilers/profiler_options/test_float_options.py
@@ -1,11 +1,12 @@
-from dataprofiler.profilers.profiler_options import IntOptions
-from dataprofiler.tests.profilers.test_numerical_options import TestNumericalOptions
+from dataprofiler.profilers.profiler_options import FloatOptions
+from dataprofiler.tests.profilers.profiler_options.test_numerical_options import TestNumericalOptions
 
 
-class TestIntOptions(TestNumericalOptions):
+class TestFloatOptions(TestNumericalOptions):
     
-    option_class = IntOptions
-
+    option_class = FloatOptions    
+    keys = TestNumericalOptions.keys + ["precision"] 
+        
     def test_init(self):
         super().test_init()
