+++ dataprofiler/tests/profilers/profiler_options/test_int_options.py
@@ -1,11 +1,10 @@
-from dataprofiler.profilers.profiler_options import TextOptions
-from dataprofiler.tests.profilers.test_numerical_options import TestNumericalOptions
+from dataprofiler.profilers.profiler_options import IntOptions
+from dataprofiler.tests.profilers.profiler_options.test_numerical_options import TestNumericalOptions
 
 
-class TestTextOptions(TestNumericalOptions):
+class TestIntOptions(TestNumericalOptions):
     
-    option_class = TextOptions
-    keys = TestNumericalOptions.keys + ["vocab"]
+    option_class = IntOptions
 
     def test_init(self):
         super().test_init()
