+++ dataprofiler/tests/profilers/profiler_options/test_profiler_class_options.py
@@ -1,5 +1,5 @@
 from dataprofiler.profilers.profiler_options import ProfilerOptions
-from dataprofiler.tests.profilers.test_base_option import TestBaseOption
+from dataprofiler.tests.profilers.profiler_options.test_base_option import TestBaseOption
 
 
 class TestProfilerOptions(TestBaseOption):
