+++ dataprofiler/tests/profilers/profiler_options/test_text_options.py
@@ -1,12 +1,12 @@
-from dataprofiler.profilers.profiler_options import FloatOptions
-from dataprofiler.tests.profilers.test_numerical_options import TestNumericalOptions
+from dataprofiler.profilers.profiler_options import TextOptions
+from dataprofiler.tests.profilers.profiler_options.test_numerical_options import TestNumericalOptions
 
 
-class TestFloatOptions(TestNumericalOptions):
+class TestTextOptions(TestNumericalOptions):
     
-    option_class = FloatOptions    
-    keys = TestNumericalOptions.keys + ["precision"] 
-        
+    option_class = TextOptions
+    keys = TestNumericalOptions.keys + ["vocab"]
+
     def test_init(self):
         super().test_init()
